+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # grinder.dart changes
 
+## UNRELEASED
+- Auto-expand directories passed to `Analyzer.analyze()` to a list of Dart 
+  source files in 
+- `coerceToPathList()` to accept `Iterable` instead of `List` and also now, in 
+  addition to `File`) converts `FileSystemEntity` (especially `Directory`) to a 
+  path.
+- Add tests for `Analyzer.analyze()` task.
+- Add defaultSourceDirectories to return the directories usually containing 
+  Dart source files.
+
 ## 0.7.1+3 (2015/5/24)
 - Fix bug with Dart.run/runAsync. Args were swapped.
 
@@ -8,7 +18,7 @@
 
 ## 0.7.1+1 (2015/5/19)
 - Add a `files` param to `TestRunner.test`.
-
+                                                                
 ## 0.7.1 (2015/5/19)
 - Added `grind --completion install` command to install tab-completion for task names
   and options in `grind`.